begin the separation
